refactor(mito2): keep compaction gate in test module
