+++ src/mito2/src/engine/compaction_test.rs
@@ -15,9 +15,11 @@
 use std::collections::HashMap;
 use std::ops::Range;
 use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Duration;
 
 use api::v1::{ColumnSchema, Rows};
+use async_trait::async_trait;
 use common_error::ext::ErrorExt;
 use common_error::status_code::StatusCode;
 use common_recordbatch::{RecordBatches, SendableRecordBatchStream};
@@ -32,11 +34,11 @@ use store_api::region_request::{
     RegionTruncateRequest, SetRegionOption, StagingPartitionDirective,
 };
 use store_api::storage::{RegionId, ScanRequest};
-use tokio::sync::Notify;
+use tokio::sync::{Notify, Semaphore};
 
 use crate::config::MitoConfig;
 use crate::engine::MitoEngine;
-use crate::engine::listener::{CompactionListener, CompactionPlanningGate};
+use crate::engine::listener::{CompactionListener, EventListener};
 use crate::test_util::{
     CreateRequestBuilder, TestEnv, build_rows_for_key, column_metadata_to_column_schema, put_rows,
 };
@@ -149,6 +151,234 @@ impl Drop for CompactionListenerGuard {
     fn drop(&mut self) {
         if let Some(listener) = self.0.take() {
             listener.wake();
+        }
+    }
+}
+
+/// Test gate that blocks selected compaction lifecycle phases for one region.
+struct CompactionPlanningGate {
+    region_id: RegionId,
+    armed: AtomicBool,
+    entered: Notify,
+    cancel_requested: Notify,
+    permits: Semaphore,
+    merge_armed: AtomicBool,
+    merge_entered: Notify,
+    merge_permits: Semaphore,
+    commit_armed: AtomicBool,
+    commit_entered: Notify,
+    commit_permits: Semaphore,
+    pending_ddl_armed: AtomicBool,
+    pending_ddl_entered: Notify,
+    pending_ddl_permits: Semaphore,
+}
+
+/// Releases an armed [`CompactionPlanningGate`] when a test exits unexpectedly.
+struct CompactionPlanningGateGuard {
+    gate: Option<Arc<CompactionPlanningGate>>,
+}
+
+impl CompactionPlanningGateGuard {
+    fn release(mut self) {
+        self.gate.take().unwrap().release();
+    }
+}
+
+impl Drop for CompactionPlanningGateGuard {
+    fn drop(&mut self) {
+        if let Some(gate) = self.gate.take() {
+            gate.release();
+        }
+    }
+}
+
+/// Releases an armed merge gate when a test exits unexpectedly.
+struct CompactionMergeGateGuard {
+    gate: Option<Arc<CompactionPlanningGate>>,
+}
+
+impl CompactionMergeGateGuard {
+    fn release(mut self) {
+        self.gate.take().unwrap().release_merge();
+    }
+}
+
+impl Drop for CompactionMergeGateGuard {
+    fn drop(&mut self) {
+        if let Some(gate) = self.gate.take() {
+            gate.release_merge();
+        }
+    }
+}
+
+/// Releases an armed commit gate when a test exits unexpectedly.
+struct CompactionCommitGateGuard {
+    gate: Option<Arc<CompactionPlanningGate>>,
+}
+
+impl CompactionCommitGateGuard {
+    fn release(mut self) {
+        self.gate.take().unwrap().release_commit();
+    }
+}
+
+impl Drop for CompactionCommitGateGuard {
+    fn drop(&mut self) {
+        if let Some(gate) = self.gate.take() {
+            gate.release_commit();
+        }
+    }
+}
+
+/// Releases an armed pending-DDL dispatch gate when a test exits unexpectedly.
+struct CompactionPendingDdlGateGuard {
+    gate: Option<Arc<CompactionPlanningGate>>,
+}
+
+impl CompactionPendingDdlGateGuard {
+    fn release(mut self) {
+        self.gate.take().unwrap().release_pending_ddl_dispatch();
+    }
+}
+
+impl Drop for CompactionPendingDdlGateGuard {
+    fn drop(&mut self) {
+        if let Some(gate) = self.gate.take() {
+            gate.release_pending_ddl_dispatch();
+        }
+    }
+}
+
+impl CompactionPlanningGate {
+    fn new(region_id: RegionId) -> Self {
+        Self {
+            region_id,
+            armed: AtomicBool::new(false),
+            entered: Notify::new(),
+            cancel_requested: Notify::new(),
+            permits: Semaphore::new(0),
+            merge_armed: AtomicBool::new(false),
+            merge_entered: Notify::new(),
+            merge_permits: Semaphore::new(0),
+            commit_armed: AtomicBool::new(false),
+            commit_entered: Notify::new(),
+            commit_permits: Semaphore::new(0),
+            pending_ddl_armed: AtomicBool::new(false),
+            pending_ddl_entered: Notify::new(),
+            pending_ddl_permits: Semaphore::new(0),
+        }
+    }
+
+    fn arm(self: &Arc<Self>) -> CompactionPlanningGateGuard {
+        self.armed.store(true, Ordering::Relaxed);
+        CompactionPlanningGateGuard {
+            gate: Some(self.clone()),
+        }
+    }
+
+    async fn wait_until_entered(&self) {
+        self.entered.notified().await;
+    }
+
+    async fn wait_until_cancel_requested(&self) {
+        self.cancel_requested.notified().await;
+    }
+
+    fn arm_merge(self: &Arc<Self>) -> CompactionMergeGateGuard {
+        self.merge_armed.store(true, Ordering::Relaxed);
+        CompactionMergeGateGuard {
+            gate: Some(self.clone()),
+        }
+    }
+
+    async fn wait_until_merge_entered(&self) {
+        self.merge_entered.notified().await;
+    }
+
+    fn arm_commit(self: &Arc<Self>) -> CompactionCommitGateGuard {
+        self.commit_armed.store(true, Ordering::Relaxed);
+        CompactionCommitGateGuard {
+            gate: Some(self.clone()),
+        }
+    }
+
+    async fn wait_until_commit_entered(&self) {
+        self.commit_entered.notified().await;
+    }
+
+    fn arm_pending_ddl_dispatch(self: &Arc<Self>) -> CompactionPendingDdlGateGuard {
+        self.pending_ddl_armed.store(true, Ordering::Relaxed);
+        CompactionPendingDdlGateGuard {
+            gate: Some(self.clone()),
+        }
+    }
+
+    async fn wait_until_pending_ddl_dispatch(&self) {
+        self.pending_ddl_entered.notified().await;
+    }
+
+    fn release(&self) {
+        self.permits.add_permits(1);
+    }
+
+    fn release_merge(&self) {
+        self.merge_permits.add_permits(1);
+    }
+
+    fn release_commit(&self) {
+        self.commit_permits.add_permits(1);
+    }
+
+    fn release_pending_ddl_dispatch(&self) {
+        self.pending_ddl_permits.add_permits(1);
+    }
+}
+
+#[async_trait]
+impl EventListener for CompactionPlanningGate {
+    async fn on_compaction_pick_begin(&self, region_id: RegionId) {
+        if region_id != self.region_id {
+            return;
+        }
+
+        if !self.armed.swap(false, Ordering::Relaxed) {
+            return;
+        }
+
+        self.entered.notify_one();
+        self.permits.acquire().await.unwrap().forget();
+    }
+
+    async fn on_merge_ssts_finished(&self, region_id: RegionId) {
+        if region_id != self.region_id || !self.merge_armed.swap(false, Ordering::Relaxed) {
+            return;
+        }
+
+        self.merge_entered.notify_one();
+        self.merge_permits.acquire().await.unwrap().forget();
+    }
+
+    async fn on_compaction_commit_begin(&self, region_id: RegionId) {
+        if region_id != self.region_id || !self.commit_armed.swap(false, Ordering::Relaxed) {
+            return;
+        }
+
+        self.commit_entered.notify_one();
+        self.commit_permits.acquire().await.unwrap().forget();
+    }
+
+    async fn on_compaction_result_notified(&self, region_id: RegionId) {
+        if region_id != self.region_id || !self.pending_ddl_armed.swap(false, Ordering::Relaxed) {
+            return;
+        }
+
+        self.pending_ddl_entered.notify_one();
+        self.pending_ddl_permits.acquire().await.unwrap().forget();
+    }
+
+    fn on_compaction_cancel_requested(&self, region_id: RegionId) {
+        if region_id == self.region_id {
+            self.cancel_requested.notify_one();
         }
     }
 }
